refactor: 메소드 네이밍 변경
